Remove non-specific IndexError
Previously there was a `tree_prefix is empty!` error being printed in a couple
of situations. 1) at startup 2) when hiding a tree group with multiple tabs in
it.

The main problem with the existing code is that it was catching a generic
IndexError without having a strong indication of what the actual cause of this
error was. We get called from `update_tab_titles()` in the base TabWidget with
a valid tab index. If we don't have information for that tab index in our tree
structure that could be an indication of a real error and we should go to
some effort to identify and prevent the cause.

This commit addresses cause 1 (at startup) by checking if the tree is empty
and returning early if so. Ideally we would have an even more specific
indication of being in startup (eg self.initialised == False) but we don't
have have a variable indicating a startup is in progress currently (we
do have one for shutting down!).

TODO: figure out when the tree is constructed vs when the tab widgets are
inserted and figure out if we can construct the tree earlier.

The second part of the commit identifies and documents the cause of 2
(collapsing a group) but leaves the error messages in for now because I want
to do a more proper fix that would let us continue to identify this
situation if we got here via another call path in the future.
This error was only being printed when hiding a group, not showing it,
because when showing a group there was more entries in the tree than in
the widget. So the function was still returning wrong data, it just
wasn't being verified at all.

The two remaining error messages are using a {foo=} f-string feature that I
don't think is available in all the python versions we support.