Avoid tab title updates while shutting down.
TreeTabWidget.get_tab_fields() was getting called with indexes that didn't
match what it had in its internal data structure on shutdown and was printing
errors. With this change tab titles no longer get updated during shutdown,
avoiding those errors.

Unfortunately we couldn't connect to the `shutting_down` signal on the parent
object because that is only emitted after the tabs are removed, and it seems
it needs to stay that way for the sake of the window close undo functionality.
So I plumbed that parent object through to a member variable, hopefully mypy
is happy with that. For some reason `self.parent()` in that later on function
was returning a MainWindow instead of a TabbedBrowser? Weird.

That does beg the question of why the nodes aren't getting removed from the
tree when they are getting removed from the widget. Should
`TreeTabbedBrowser._remove_tab()` be doing things in a different order? That
says "node will already be removed from tree", so where are they getting
removed? Or is that meaning widget instead of tree?

TODO: get more certainty on alternate logic fixes