Remove tab from parent before we remove it from the tree.
This resolves an issue in update_tab_titles() when deleting a tab with
collapsed children. Since we were doing all of the super() action before
removing the node from the tree the super() method was calling
on_current_changed() before we had done our bit of updating the tree.
update_tab_titles() was getting called from that which was calling into
get_tab_fields() in TreeTabWidget which was complaining that the tree and
widget didn't match up.

We don't have any automated tests covering this behaviour currently but moving
the super() call to the end works so far. Possibly it was that way due to some
earlier refactoring.

I've changed a local variable name since it was shadowing a method arg that is
now used further down.

TODO:
* are there other pieces of code where we should look whether we are doing
  tree stuff or widget stuff first and enshrine that as a pattern?
* the note about `_add_undo_entry()` points to unfortunate code that we should
  look into. It doesn't seem correct to be manipulating the tree in that
  method and the explanatory comment is too vague to explain anything