Switch a couple more traversals to child first
I think when removing nodes it's less work to remove leaves first, so that
children don't have to be reparented. We also don't need to save everything to
a list. I don't know what the difference out of POST and POST_R, maybe if you
remove the last sibling first it's less work? Not sure.

Also take care of a fixme comment which is fixed (the comment it refers to is
still there, but add_undo_entry is no longer modifying the tree).