Check the correct tab is pinned when closing
When recursive closing a tree group you will be prompted if you want to
close any tabs that are pinned. Except we passed the topmost tab to the
prompt method every time instead of the tab we were going to close.
This fixes that. It does mean that if there are multiple pinned tabs in
a tree it gives you an identical prompt for each one, which no
indication that they are separate prompts. It looks like nothing is
happening when you answer it. But that's a separate problem (probably
the solution is to either prompt once or to put URLs and/or tab IDs in
the prompt).