tests: Ignore yet another mesa (?) warning
As shown in detail by toofar in dff25d10bcc8a398a77ef260c33e51ec5c70f482,
mesa 23.3 seems to trigger some new error messages.

On my setup locally, it also triggers the warning that's ignored here.

Why that's not the case on CI or for others is unclear to me, I gave up
trying to understand graphics stuff on Linux.

I'll just presume this is the same underlying cause, and refer to the
commit message of the commit above for more details.