Support loading legacy format tree tab sessions.
In a2f98aa78b52 the session format for tree tab windows was changed to be
backwards compatible with the prior format. After that change we can load
non-tree tab windows fine but the browser crashes on startup when trying to
load session in the now-legacy tree tab session format.

Since people have been using the now-legacy session for years and there is no
way for them to easily migrate I figure we have a few options:

* provide a script to help them migrate
* handle the error encountered (`KeyError: 'tabs`) when trying to load a
  legacy session better, eg give them a nice apologetic message
* support loading the old style of sessions

It didn't take much effort to support the prior format, so I propose we do
that. The data is all the same, we just changed it to nodes are childs of tabs
instead of the other way around.