Support modifying suffixes in "Given I open ..."
Previously only "I open" in the "When" (aka "act") block supported suffixes
like "in a new windows". The tree tabs setting only applies to new windows so
I want to open a new window after setting the setting.

So I changed the given `open_path` to call the other `open_path`, then had to
support the difference between them: the Given version always opens the path
in a new tab. I changed that to open in a new tab *by default*.

It was a bit tricky that just eg passing in a `default_new_tab` arg and then
passing `new_tab=True` to quteproc command because then if I ended up passing
`new_window` based on a suffix it rightly complained that those two args
conflicting. So I only really want to set `new_tab=True` if no suffixes are
detected. Instead of adding a new `suffix_found = True` line to every part of
the if chain I converted it to loop over a dict.

I didn't think too hard about the flexible typing of the values of the suffix
map. Just did something for every case I observed. It probably works fine.