Inch a little closer towards compare tree tabs session
Had a think about how to get tree structure things to compare. For now
I'm thinking we create a string of the format:

    - about:blank?one
      - about:blank?child (collapsed)
        - about:blank?grandchild
      - about:blank?childtwo (active)
    - about:blank?two

From both a) the text fixture (we write it like that) and b) the session
data (generate it). Then we just compare the strings, hopefully much
like check_open_tabs and maybe we can even make one implementation work
for both.

The next bit to do is generate the above text from the session data. We
can start by loading the tree structure with
`SessionManager._reconstruct_tree_data()` (TODO: make that public) and
then we need to recurse through that. We can take inspiration from
`recursive_load_node()` in sessions.py

Then copy the rest of the implementation from `check_open_tabs()` and
adapt that.

And then look at `quteproc.compare_session()` as well I guess. But it
looks like it'll be a pain to write out the whole tree tab format
sessions...