Fix tab widget benchmark tests
In 8aa88b83e I added a dependency in the TabWidget in
`parent.is_shutting_down` (assuming parent is a TabbedBrowser, but
without updating the type hints...).

Since the benchmark test wasn't setting a parent it's been breaking
since then, oops.

That's going to be breaking all the branches based off of the tree tabs
branch, so we should probably cherry pick this back to the
tree-tabs-integration branch. And consider updating the
TabWidget parent kwarg type hint (and make the dummpy parent a
mock/spy/double of the TabbedBrowser I guess).

Not sure where the size of 640x480 came from before hand but when I
added a parent it went to 100x30.