Be explicit about tab position when loading from sessions
With, presumably, rare combinations of settings tab order in windows
loaded from session can be different from the window the session was
saved from.

In particular with `tabs.new_position.related=prev` the tabs would be
loaded in reverse order!

This commit changes tabs loaded from sessions to 1) override positioning
related kwargs to tabopen so that it doesn't look at user sessions 2)
provide an explicit index to use for new tabs.

This particular test passes with either one of the newly passed kwargs.
We don't strictly need both of them, but while being explicit we may as
well override everything to be extra clear.

While looking at this code you may be asking why background is set to
True for all the tabs when only one of them is going to be focused? Good
question, I don't think it should be, see https://github.com/qutebrowser/qutebrowser/issues/6983#issuecomment-1025011641

But since that is explicitly set it should probably be a separate PR to
change it. Looks like the original PR to fix that got caught up in wider
scope and trying to do a new style of e2e-but-python tests. It should be
easy enough to write an e2e that loads a page with JS that reports that
visibility state.

TODO:
* add a similar test for tree tabs when the scaffolding is there (with
  `new_position.tree.new_toplevel=prev` too)
* override sibling too?