Fix another possible tree and widget desync while loading a session
In 3ec2000b2f1c I identified a few cases where we could be asked to call into
the tree structure based on a widget index with evidence that the tree and
widget where out of sync. In particular where there were a different amount of
items in each.

Previously I had ignored desyncs during session loads but now I've found
a case where the desync remains after the session was loaded, for a
short time. It can be reproduced (most of the time) with:

   python3 -m qutebrowser -T -s tabs.tree_tabs true ":open about:blank?1" ":open -tr about:blank?2" ":open -tr about:blank?3" ":tab-focus 2" ":tree-tab-toggle-hide" ":cmd-later 200 session-save foo" ":cmd-later 300 session-load -c foo"

This only happens if the last set of tabs we open is a collapsed group. There
would have been no update event fired since we set the collapsed attribute.

If this issue shows up again a more robust approach might be to not set the
collapsed attribute in the loop, but load all the tabs visible then go and set
the collapsed attribute on all the nodes that need it and fire an update.

I initially reproduced this with the "Load a collapsed subtree" end2end test.