Wait for all tabs to be loaded after loading sessions
I got a failure in CI on windows for the "Load a collapsed subtree"
where it got an index error when trying to load the active history entry
for a tab in a session. This is likely because the session was saved
before the tab was loaded. I can reproduce it locally by changing the
last test to wait for only the first tab, instead of the last.