Remove unneeded `_init_config()` override from TreeTabWidget
The method in this function where added in: 411d39a2e6e "Add config
option to enable tree-tabs". The `update_tab_titles()` line has the
comment "must also be called when deactivating" because back then
it was right before something gated on `if tabs.tree_tabs:`. But:

    1. tree_tab_update() already calls update_tab_titles()
    2. we don't currently support toggling tree tab off and on on an existing
       window

So I don't believe the calls in this override are needed with the current
design. `_init_config()` is called on `__init__()` and `tabs.*` config
items changing. To double check if these is needed for anything I've
tried a bunch of things like:

    1. change tab format
    2. show/hide tab group
    3. change tabs.position
    4. show/hide tab bar - huh, when the tab bar is re-shown it's scrolled
       down so on the title of the active tab and below are shown, is this
       new? But that happens with a flat tabs window too. It only happens
       sometimes though. Oh, it's not happening on 6.4.2 but is on 6.6.0 and I
       don't remember it happending on 6.7.0. Must be a Qt issue that
       hopefully got fixed.

And it all seemed to behave normally. I'm not sure what else to check
as everything else tree tab specific that I can think of relates to
moving tabs, which should be updating things already.