Remove unneeded `__init__` override from TreeTabbedBrowser
The only thing that the overridden init was doing was creating a
different widget then duplicating the signal setup. I've moved the
widget creation out to a helper function so that we only need to
override that. This has the benefit of not needlessly creating the
non-tree TabWidget.

An alternative to moving the tab widget out to a helper would be to add
a `_set_tab_widget(widget)` method that did the signal connecting and
then keep the widget creating in init and pass the widget to the setter.
But then 1) there would be a bigger diff 2) I would have to think about
what to do if a widget was already set, complain? do nothing? disconnect
and dispose of the previous one?