Remove `treetabbedbrowser._add_undo_entry()`
This one's not as clear a win as the previous one. The main duplication
here was manipulating `undo_stack`. There was very little in
`_add_undo_entry()` apart from that, and it was already being done in
the parent. Sadly there isn't really a win in terms of line count (there
is slightly more lines now but that's because of comments and formatting
changes).

I pulled out a ` _add_undo_entry()` method in the parent class, even
though it isn't overriden anymore, just because it seemed like a good
idea and makes the `_remove_tab()` method a bit cleaner.

I moved the tree tab specific logic (collapsed tabs and parent and child
data) into the UndoEntry class because I don't really see an alternative
to having a specialised one of them, so may as well make use of it. Also
using it as a dumb data class and having the related class contain logic
related to it _pushes up glasses_ violates one or two OOP principles.

Having both undo classes handle a `browsertab.WebTabError` is a minor
bit of duplication but if the parent class handled it I think it's too
non-specific an error to be passing up to levels (it would seem a bit
"magical" how the parent class inferred the meaning of it) and adding a
`history_data` arg into `from_tab()` seems like pointless toil for the
parent class. There is probably more "pointless toil" that could be
removed from the parent class and pulled into the specialised class that
knows about undo logic, but that isn't the goal of this PR.

I had to teach the parent class about `_undo_class.from_tab()` possibly
returning a list even though it would never do that for a flat
TabbedBrowser, but oh well. That nested conditional could maybe be
arguably optimised but I'm not seeing a clear win.

Regarding next steps around here, I think
`TreeTabbedBrowser._remove_tab()` is the bigger win than the undo logic
itself. It entirely deals with tree stuff which I feel should be owned
by the tree data structure, and moving that there would probably mean
other pieces of code could be switched to use it too. Eg the
re-parenting could be in the notree class and if we taught the notree
nodes how to close tabs the collapsed node handling could be moved there
too.